Fix sidebar scrim review feedback
